LayoutTests/fast/security-labels/submit-event-expected.txt

CONSOLE MESSAGE: Not allowed to submit form by checksubmit: http://localhost/foo
CONSOLE MESSAGE: Not allowed to submit form by checksubmit: http://localhost/foo
This tests that the checksubmit event behaves properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


checksubmit reached with http://localhost/foo
PASS tag.isOn(event.securityLabel) is false
checksubmit reached with http://localhost/foo
PASS tag.isOn(event.securityLabel) is true
PASS successfullyParsed is true

TEST COMPLETE
